// frontend/src/lib/proto/masters/v1/master_pb.d.ts
export declare type ProductCategory = Message<"masters.v1.ProductCategory"> & {
  /**
   * @generated from field: string id = 1;
   */
  id: string;

  /**
   * @generated from field: string company_id = 2;
   */
  companyId: string;

  /**
   * @generated from field: string department_id = 3;
   */
  departmentId: string;

  /**
   * @generated from field: string code = 4;
   */
  code: string;

  /**
   * @generated from field: string name = 5;
   */
  name: string;

  /**
   * @generated from field: int32 sort_order = 6;
   */
  sortOrder: number;

  /**
   * @generated from field: bool is_active = 7;
   */
  isActive: boolean;

  /**
   * @generated from field: string created_at = 8;
   */
  createdAt: string;

  /**
   * @generated from field: string updated_at = 9;
   */
  updatedAt: string;

  /**
   * @generated from field: string deleted_at = 10;
   */
  deletedAt: string;
};